% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adpRemoveEmptyBins.R
\name{adpRemoveEmptyBins}
\alias{adpRemoveEmptyBins}
\title{Remove adp bins that contain all NA values}
\usage{
adpRemoveEmptyBins(adp, debug = 1, fields = NULL)
}
\arguments{
\item{adp}{an \linkS4class{adp} object containing bottom ranges.}

\item{debug}{integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information.}

\item{fields}{a variable contained within \code{adp} indicating which field to flag.
If NULL (the default) then \code{\link[=adpRemoveEmptyBins]{adpRemoveEmptyBins()}} applies itself to all flag
fields that have the same dimensions as \code{v} in the \code{data} slot.}
}
\value{
\code{adpRemoveEmptyBins} returns an \linkS4class{adp} object only
distances with finite values
}
\description{
This function removes distance bins that have been flagged as bad,
likely from \link[oce:adpFlagPastBoundary]{oce::adpFlagPastBoundary} and then set to NA, likely from
\link[oce:handleFlags]{oce::handleFlags}. It does this by identifying the minimum bin that contains
all NA values. If this minimum bin is found to be within the upper 30 m,
and likely a result of manual flagging due to another instrument on the moored
line, it will skip that and move to the next one. The fields with the
same dimensions as "v", as well as the "distance" variable, and "flags"
are then replaced to only contain the distance bins with at least one finite value.
}
\examples{
\dontrun{
library(oce)
library(oceToNetCDF)
load(system.file("extdata", "adcpNA", package="oceToNetCDF"))
adp2 <- adpFlagPastBoundary(adcpNA)
adp3 <- adpConvertRawToNumeric(adp2)
adp4 <- handleFlags(adp3, flags=4)
adp5 <- adpRemoveEmptyBins(adp4, debug=1)
par(mfrow=c(2,1))
plot(adp4, which="a1")
plot(adp5, which="a1")
}

}
\author{
Jaimie Harbin
}
