% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertConductivityRatio.R
\name{convertConductivityRatio}
\alias{convertConductivityRatio}
\title{Convert conductivity ratio to conductivity}
\usage{
convertConductivityRatio(crat, unit = "S/m")
}
\arguments{
\item{crat}{a conductivity ratio number}

\item{unit}{the specified unit of conductivity to convert to
(Either 'S/m' or 'mS/cm')}
}
\value{
a conductivity value in the specified unit
}
\description{
Convert conductivity ratio to conductivity
}
