% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixMetadata.R
\name{fixMetadata}
\alias{fixMetadata}
\title{Clean metadata for a CTD and RCM object}
\usage{
fixMetadata(odf, debug = 0, data = NULL, unit = NULL)
}
\arguments{
\item{odf}{an odf object (oce::read.odf())}

\item{debug}{integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information.
@importFrom oce}

\item{data}{a data frame of standard name, name, units, and GF3 codes likely from getData}

\item{unit}{standardized unit S/m by default}
}
\value{
an odf object with proper units and flags
}
\description{
This function ensures the proper units are associated with each variable
and adds place holders for flags if they do not already exist.
}
