% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStandardData.R
\name{getStandardData}
\alias{getStandardData}
\title{Get required data for specified instrument type}
\usage{
getStandardData(type = "ctd", standard = "cf", silent = TRUE)
}
\arguments{
\item{type}{the type of data that will eventually be converted into a netCDF.
The options are ctd, rcm, or adcp}

\item{standard}{parameter equal to either 'cf' (Climate or Forecast) or "bodc"
British Oceanographic Data Center) indicating which standard data to obtain}

\item{silent}{logical value indicating whether to silence some CF compliant warnings. The default is to silent such indicators.}
}
\value{
a data frame containing the necessary standard names, units, codes, names, and type of data
}
\description{
This function gets the available standard names, units, codes,names, and
type of data for the required information for the specified instrument type
of either ctd, rcm, or adcp.
}
\examples{
data <- getStandardData(type="ctd", standard="cf")
names(data)
}
