% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlabToOce.R
\name{matlabToOce}
\alias{matlabToOce}
\title{Create ctd objects from Matlab CTD file with finite data}
\usage{
matlabToOce(matfile = NULL, debug = 0)
}
\arguments{
\item{matfile}{a moored CTD matlab file}

\item{debug}{integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information.}
}
\value{
a list of oce object with finite data values
from multiple stations
}
\description{
This function removes NaN values in moored CTD Matlab files.
It starts by reading a matlab file \code{\link[R.matlab:readMat]{R.matlab::readMat()}}, then
cycling through the multiple profiles of multiple stations
to determine which data values are finite. The finite data
values are then used to create an oce ctd object using
\code{\link[oce:as.ctd]{oce::as.ctd()}}.
}
\examples{
\dontrun{
library(oce)
library(R.matlab)
filesInst <- list.files(path="./20_rawDavis/instData/", full.names = TRUE)
x <- filesInst[[2]]
ctds <- matlabToOce(x)
}
}
