% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameReplacement.R
\name{nameReplacement}
\alias{nameReplacement}
\title{Replace DFO Codes with CF standard names}
\usage{
nameReplacement(odf, data = NULL, debug = 0)
}
\arguments{
\item{odf}{an odf object (oce::read.odf())}

\item{data}{a data frame of standard name, name, units, and GF3 codes likely from getData}

\item{debug}{integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information.}
}
\value{
an odf file with the dataOringinalNames as standard CF
names and the institute as UW/DFO
}
\description{
Replace DFO Codes with CF standard names
}
\examples{
library(odfToNetCDF)
library(oce)
data <- getData(type="ctd")
f <- system.file("extdata", "mctd.ODF", package="odfToNetCDF")
odf1 <- read.odf(f)
odf2 <- nameReplacement(odf1, data=data)
}
