% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odfMetadataCheck.R
\name{odfMetadataCheck}
\alias{odfMetadataCheck}
\title{Check if required metadata exists}
\usage{
odfMetadataCheck(odf, data = NULL, print = TRUE, debug = 0)
}
\arguments{
\item{odf}{an odf object obtained from \code{\link[oce:read.odf]{oce::read.odf()}}}

\item{data}{a data frame of standard name, name, units, and GF3 codes likely from getStandardData}

\item{print}{TRUE or FALSE, TRUE will cause errors to be displayed at command
line, FALSE will sink errors into text document}

\item{debug}{integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information}
}
\value{
a print statement of any issues with metadata to command line, if nothing
prints then all metadata is intact
}
\description{
This function checks that an ODF object has all the metadata
required to build a complete netCDF file
}
\examples{
library(odfToNetCDF)
library(oce)
data <- getStandardData(type="ctd")
f <- system.file("extdata", "mctd.ODF", package="odfToNetCDF")
odf1 <- read.odf(f)
odfMetadataCheck(odf1, data=data)
}
