% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{package}
\name{odfToNetCdf-package}
\alias{odfToNetCdf-package}
\title{A Package for converting ODF files to NetCDF files}
\description{
The \code{odfToNetcdf} package provides tools for converting moored conductivity-temperature-
depth ("CTD"), acoustic-doppler-current-profiler (ADCP), and current meter (CM) ODF
files into NetCDF files.
}
\details{
The sketches given below illustrates the typical workflow with the package, with
descriptions of the steps on the left, and names of the relevant functions
on the right.
}
