% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polishODF.R
\name{polishODF}
\alias{polishODF}
\title{Clean data and metadata for a CTD and RCM object}
\usage{
polishODF(odf, debug = 0, data = NULL)
}
\arguments{
\item{odf}{an odf object (oce::read.odf())}

\item{debug}{integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information.}

\item{data}{a data frame of standard name, name, units, and GF3 codes likely from getData}
}
\value{
an odf object
}
\description{
This function ensures the proper units are associated with each variable
and adds place holders for flags if they do not already exist.
For a CTD type, this function also converts conductivity ratio to
conducitivity if the proper unit is provided to abide by CF standards.
}
\examples{
library(odfToNetCDF)
library(oce)
data <- getData(type="ctd")
f <- system.file("extdata", "mctd.ODF", package="odfToNetCDF")
odf1 <- read.odf(f)
odf2 <- nameReplacement(odf1, data=data, unit="S/m")
odf3 <- removeDerived(odf2)
odf4 <- polishODF(odf3, data=data)
}
