% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDerived.R
\name{removeDerived}
\alias{removeDerived}
\title{Remove derived ctd data and metadata}
\usage{
removeDerived(odf, debug = 0)
}
\arguments{
\item{odf}{an odf object (oce::read.odf())}

\item{debug}{integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information}
}
\value{
CTD object: an odf object with only data for time, conductivity, salinity,
temperature, and pressure. RCM object: an odf object with only data for horizontal_current_direction,
barotropic_sea_water_x_velocity, sea_water_pressure, sea_water_practical_salinity", "time",
and sea_water_temperature
}
\description{
This function removes data and metadata that is derived.
For a CTD type, the only data and metadata kept is time, conductivity, salinity,
temperature, and pressure. For an RCM type, the only data and
metadata kept is horizontal_current_direction,
barotropic_sea_water_x_velocity, sea_water_pressure, sea_water_practical_salinity,
time, and sea_water_temperature
}
\examples{
library(odfToNetCDF)
library(oce)
data <- getData(type="ctd")
f <- system.file("extdata", "mctd.ODF", package="odfToNetCDF")
odf1 <- read.odf(f)
odf2 <- nameReplacement(odf1, data=data, unit="S/m")
odf3 <- removeDerived(odf2)
}
