% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardMetadata.R
\name{standardMetadata}
\alias{standardMetadata}
\title{Insert Metadata names to be compliant with IOOS}
\usage{
standardMetadata(x, fixed = FALSE, file = NULL)
}
\arguments{
\item{x}{an \linkS4class{oce} object containing bottom ranges.}

\item{fixed}{a Boolean indicating if the answers input by the user
should be saved as a data frame and applied to a list of \code{oce} objects
See Details.}

\item{file}{a file name indicating the name to save a data frame if \code{fixed=TRUE}.
If this file name already exists in the present directory, the answers from this
data frame are used.}
}
\value{
`an \linkS4class{oce} object containing highly reccommended IOOS metadata
}
\description{
This function inserts the required global attributes that are
"highly recommended" for the IOOS (Integrated Ocean Observing
System) standards (see Reference 1). This function works by
extracting what it can from the data and metadata within the
\code{oce} object, but if its required field is determined to be NULL
the user will be prompted to enter in the required information.
If \code{fixed=TRUE}, the information that is input by the user will
be saved and applied to a list of \code{oce} objects. See \sQuote{Details}.
}
\details{
If \code{fixed = TRUE} and a data frame named with the specified file name
doesn't exist it will save the user's answers into a data frame named
\code{file} which contains the name of the field that required a users answers as well
as the answers the user input. This means, if \code{fixed=TRUE} and a file named \code{file}
does exist, this function will load up the previously answered questions
to avoid the user needing to answer them again.
}
\examples{
\dontrun{
library(oceToNetCDF)
library(oce)
data <- getStandardData(type="ctd")
f <- system.file("extdata", "mctd.ODF", package="oceToNetCDF")
odf1 <- read.odf(f)
odf2 <- standardMetadata(odf1)
odf3 <- nameReplacement(odf2, data=data, unit="S/m")
odf4 <- removeDerived(odf3)
odf5 <- fixMetadata(odf4, data=data)
}

}
\references{
\enumerate{
\item https://wiki.esipfed.org/Attribute_Convention_for_Data_Discovery_1-3#Global_Attributes
}
}
\author{
Jaimie Harbin
}
