% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCrat.R
\name{checkCrat}
\alias{checkCrat}
\title{Check range of conductivity}
\usage{
checkCrat(odf, unit = "S/m")
}
\arguments{
\item{odf}{an odf object (oce::read.odf())}

\item{unit}{the specified unit of conductivity to convert to
(Either 'S/m' or 'mS/cm')}
}
\value{
a print statement indicating if the conductivity
is in the expected range
}
\description{
Check range of conductivity
}
