% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCrat.R
\name{checkCrat}
\alias{checkCrat}
\title{Check range of conductivity}
\usage{
checkCrat(odf, unit = "S/m")
}
\arguments{
\item{odf}{an odf object (oce::read.odf())}

\item{unit}{the specified unit of conductivity to convert to
(Either 'S/m' or 'mS/cm')}
}
\value{
a print statement indicating if the conductivity
is in the expected range
}
\description{
Check range of conductivity
}
\examples{
\dontrun{
library(odfToNetCDF)
data <- getData(type="ctd")
odf1 <- read.odf("MCTD_KN179-05_1533_3309_1800.ODF")
odf2 <- nameReplacement(odf1, data=data)
odf3 <- removeDerived(odf2)
odf4 <- polishODF(odf3, data=data, unit='S/m')
odf5 <- checkCrat(odf3)
}
}
