% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileOdfToAdp.R
\name{compileOdfToAdp}
\alias{compileOdfToAdp}
\title{Compile multiple adp ODF files to a single adp object}
\usage{
compileOdfToAdp(files)
}
\arguments{
\item{files}{list of adp odf files}
}
\description{
This function reads in a set of adp odf files and compiles them
into a single adp object. It is capable of recognizing missing bins
or missing odf files and inserts NA values in appropriate slot of
data variables.
}
\examples{
\dontrun{
library(odfToNetCDF)
files <- list.files(pattern="ODF")
adp <- compileOdfToAdp(files)
}

}
