% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileOdfToAdp.R
\name{compileOdfToAdp}
\alias{compileOdfToAdp}
\title{Compile multiple adp ODF files to a single adp object}
\usage{
compileOdfToAdp(files, metadata)
}
\arguments{
\item{files}{list of odf files}

\item{metadata}{any extra metadata to be added to net cdf as list form}
}
\description{
Converting individual odf bins to Net cdf standard format
}
\details{
This function reads in a set of odf files and compiles into adp object.
It is capable of recognizing missing bins or missing odf files and
inserting NA values in appropriate slot of data variables
}
\concept{odf}
