% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileOdfToAdp.R
\name{compileOdfToAdp}
\alias{compileOdfToAdp}
\title{Compile multiple adp ODF files to a single adp object}
\usage{
compileOdfToAdp(files, debug = 0)
}
\arguments{
\item{files}{list of adp odf files}

\item{debug}{integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information}
}
\description{
This function reads in a set of adp odf files and compiles them
into a single adp object. It is capable of recognizing missing bins
or missing odf files and inserts NA values in appropriate slot of
data variables.
}
\examples{
library(oceToNetCDF)
data <- getStandardData(type="adcp")
f1 <- system.file("extdata", "adcp1.ODF", package="oceToNetCDF")
f2 <- system.file("extdata", "adcp2.ODF", package="oceToNetCDF")
files <- c(f1,f2)
adp <- compileOdfToAdp(files)

}
