% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertNetCDF.R
\name{convertNetCDF}
\alias{convertNetCDF}
\title{Template to convert odf object to netCDF}
\usage{
convertNetCDF(odf, filename = NULL, debug = 0, data = NULL)
}
\arguments{
\item{odf}{an odf object from oce which contains mctd data}

\item{filename}{the desired name for the netCDF file produced, if left NULL
the default will conform to BIO naming conventions}

\item{debug}{integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information.}

\item{data}{a data frame of standard name, name, units, and GF3 codes likely from getData}
}
\value{
netCDF file with a maximum of 12 variables
}
\description{
This function converts and odf object to a netCDF files for CTD
and RCM types.
}
\examples{
library(odfToNetCDF)
library(oce)
data <- getData(type="ctd")
f <- system.file("extdata", "mctd.ODF", package="odfToNetCDF")
odf1 <- read.odf(f)
odf2 <- nameReplacement(odf1, data=data)
odf3 <- removeDerived(odf2)
odf4 <- polishODF(odf3, data=data, unit='S/m')
}
