% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixMetadata.R
\name{fixMetadata}
\alias{fixMetadata}
\title{Clean metadata for a ctd and rcm objects}
\usage{
fixMetadata(odf, debug = 0, data = NULL)
}
\arguments{
\item{odf}{an odf object \code{\link[oce:read.odf]{oce::read.odf()}}}

\item{debug}{integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information.}

\item{data}{a data frame of standard name, name, units, and GF3 codes likely from \code{\link[=getData]{getData()}}}
}
\value{
an odf object
}
\description{
This function ensures the proper units are associated with each variable
and adds place holders for flags if they do not already exist.
}
\examples{
library(odfToNetCDF)
library(oce)
data <- getData(type="ctd")
f <- system.file("extdata", "mctd.ODF", package="odfToNetCDF")
odf1 <- read.odf(f)
odf2 <- nameReplacement(odf1, data=data, unit="S/m")
odf3 <- removeDerived(odf2)
odf4 <- fixMetadata(odf3, data=data)
}
