% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCFData.R
\name{getCFData}
\alias{getCFData}
\title{Get required data for specified instrument type}
\usage{
getCFData(type = "ctd")
}
\arguments{
\item{type}{the type of data that will eventually be converted into a netCDF.
The options are ctd, rcm, or adcp}
}
\value{
a data frame containing the necessary standard names, units, codes, names, and type of data
}
\description{
This function gets the available standard names, units, codes,names, and
type of data for the required information for the specified instrument type
of either ctd, rcm, or adcp.
}
\examples{
data <- getCFData(type="ctd")
names(data)
}
