% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameReplacement.R
\name{nameReplacement}
\alias{nameReplacement}
\title{Replace DFO codes with CF standard names}
\usage{
nameReplacement(odf, data = NULL, debug = 0, institute = NULL, unit = NULL)
}
\arguments{
\item{odf}{an odf object \code{\link[oce:read.odf]{oce::read.odf()}}}

\item{data}{a data frame of standard name, name, units, and GF3 codes likely from \code{\link[=getCFData]{getCFData()}}}

\item{debug}{integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information.}

\item{institute}{an optional name to add as the institute name}

\item{unit}{specified unit of 'S/m' or 'mS/cm' to convert CRAT to
sea_water_electrical_conductivty for CF standard. This is only
used for ctd type}
}
\value{
an odf file
}
\description{
This function replaces DFO codes for dataNamesOriginal and data
with CF standards. For ctd types, if conductivity ratio (CRAT)
exists, the values are converted to sea_water_electrical_conductivity
values to abide by CF standards, and the unit is changed to the
specified unit.
}
\examples{
library(odfToNetCDF)
library(oce)
data <- getCFData(type="ctd")
f <- system.file("extdata", "mctd.ODF", package="odfToNetCDF")
odf1 <- read.odf(f)
odf2 <- nameReplacement(odf1, data=data, unit="S/m")
}
