% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleAdpNetCDF.R
\name{singleAdpNetCDF}
\alias{singleAdpNetCDF}
\title{Convert adp object to NetCDF}
\usage{
singleAdpNetCDF(adp, name, metadata, debug = 0, data = NULL)
}
\arguments{
\item{adp}{an adp adpect from the oce class}

\item{name}{name of the NetCDF file to be produced}

\item{metadata}{csv file listing metadata names and values to be inserted into
global attributes of net CDF}

\item{data}{a data frame of standard name, name, units, and GF3 codes likely from getData}
}
\description{
Exports an adp object to a netCDF using variables and metadata
within adp combined with optional additional metadata
}
