% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleAdpNetCDF.R
\name{singleAdpNetCDF}
\alias{singleAdpNetCDF}
\title{Convert adp object to netCDF}
\usage{
singleAdpNetCDF(adp, name, debug = 0, data = NULL)
}
\arguments{
\item{adp}{an adp adpect from the oce class}

\item{name}{name of the netCDF file to be produced}

\item{debug}{integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information.}

\item{data}{a data frame of standard name, name, units, and GF3 codes likely from getData}
}
\description{
This function exports an adp object to a netCDF using variables
and metadata within adp combined
}
\examples{
\dontrun{
library(odfToNetCDF)
library(oce)
data <- getData(type="adcp")
files <- list.files(pattern="ODF")
adp <- compileOdfToAdp(files)
singleAdpNetCDF(adp, name="test", debug=1, data=data)
}

}
