% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardName.R
\name{standardName}
\alias{standardName}
\title{Obtain standard name, name, and units from GF3 code}
\usage{
standardName(gf3, data = NULL, debug = 0)
}
\arguments{
\item{gf3}{a character indicating a GF3 (General Formatting) standard code parameter}

\item{data}{a data frame of standard name, name, units, and GF3 codes likely from getCFData}

\item{debug}{integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information.}
}
\value{
A list containing the standard name and unit of the GF3 code in CF standard
}
\description{
This function reveals the standard name, name, and units in climate and
forecast (CF) standards from a General Formatting (GF3) standard
code parameter.
}
\examples{
library(odfToNetCDF)
data <- getCFData(type="ctd")
standardName("TEMP", data=data)
}
