% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardName.R
\name{standardName}
\alias{standardName}
\title{Obtain standard name and units from GF3 code}
\usage{
standardName(gf3, data = NULL)
}
\arguments{
\item{gf3}{a character indicating a GF3 (General Formatting) standard code parameter}
}
\value{
A list containing the standard name and unit of the GF3 code in CF standard
}
\description{
This function reveals the standard name and units in climate and
forecast (CF) standards from a General Formatting (GF3) standard
code parameter.
}
